:construction: Changed remaining User class ID types from Integer to Long

+++ src/main/java/com/schnabel/schnabel/users/service/PatientService.java
@@ -7,7 +7,7 @@ import com.schnabel.schnabel.users.repository.IPatientRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class PatientService extends CrudService<Patient, Integer> implements IPatientService
+public class PatientService extends CrudService<Patient, Long> implements IPatientService
 {
 
     public PatientService(IPatientRepository patientRepository)
